[REFACTOR][RUNTIME][RPC] Move minrpc/ into src/runtime/rpc/minrpc/
minrpc implements a minimal RPC protocol that shares rpc_reference.h
with the standard RPC transport in src/runtime/rpc/. Nesting minrpc/
under rpc/ makes that structural relationship explicit on the file
system and removes it from the runtime top-level.

Changes:
- src/runtime/minrpc/ → src/runtime/rpc/minrpc/ (3 files + posix_popen_server/)
- rpc_endpoint.h, rpc_session.h: "../minrpc/rpc_reference.h" → "minrpc/rpc_reference.h"
- disco/protocol.h, process_session.cc, threaded_session.cc:
  "../minrpc/rpc_reference.h" → "../rpc/minrpc/rpc_reference.h"
- hexagon/rpc/{hexagon,simulator}/rpc_server.cc:
  "../../../minrpc/minrpc_server.h" → "../../../rpc/minrpc/minrpc_server.h"
- CMakeLists.txt: RUNTIME_SRCS glob src/runtime/minrpc/*.cc → src/runtime/rpc/minrpc/*.cc
- cmake/modules/Hexagon.cmake: two explicit header list entries updated
- python/tvm/rpc/minrpc.py: path constant for -I flag and server.cc source
  updated from src/runtime/minrpc to src/runtime/rpc/minrpc

Header guards in minrpc_server.h and rpc_reference.h already encode
TVM_RUNTIME_MINRPC_*, which remain correct after the move.
