+++ src/runtime/disco/process_session.cc
@@ -27,7 +27,7 @@
 #include <vector>
 
 #include "../../support/pipe.h"
-#include "../minrpc/rpc_reference.h"
+#include "../rpc/minrpc/rpc_reference.h"
 #include "./bcast_session.h"
 #include "./disco_worker_thread.h"
 #include "./message_queue.h"
