+++ src/runtime/hexagon/rpc/hexagon/rpc_server.cc
@@ -36,7 +36,7 @@ extern "C" {
 #include <memory>
 #include <string>
 
-#include "../../../minrpc/minrpc_server.h"
+#include "../../../rpc/minrpc/minrpc_server.h"
 #include "../../hexagon/hexagon_common.h"
 #include "../../hexagon/hexagon_device_api.h"
 #include "../../profiler/prof_utils.h"
