+++ src/runtime/rpc/rpc_endpoint.h
@@ -32,7 +32,7 @@
 #include <utility>
 
 #include "../../support/ring_buffer.h"
-#include "../minrpc/rpc_reference.h"
+#include "minrpc/rpc_reference.h"
 #include "rpc_channel.h"
 #include "rpc_session.h"
